fix: rename disallow to deny
