+++ pkg/controller/cp/controller.go
@@ -102,7 +102,7 @@ func (ctrl *Controller) readPolicy(logE *logrus.Entry, param *config.Param) ([]*
 	if policyFile != "" {
 		if err := ctrl.policyValidator.Validate(policyFile); err != nil {
 			if err := ctrl.policyValidator.Warn(logE, policyFile); err != nil {
-				logE.WithError(err).Warn("warn an disallowed policy file")
+				logE.WithError(err).Warn("warn an denied policy file")
 			}
 		} else {
 			param.PolicyConfigFilePaths = append(param.PolicyConfigFilePaths, policyFile)
