+++ pkg/controller/install/install.go
@@ -79,7 +79,7 @@ func (ctrl *Controller) Install(ctx context.Context, logE *logrus.Entry, param *
 	if policyFile != "" {
 		if err := ctrl.policyValidator.Validate(policyFile); err != nil {
 			if err := ctrl.policyValidator.Warn(logE, policyFile); err != nil {
-				logE.WithError(err).Warn("warn an disallowed policy file")
+				logE.WithError(err).Warn("warn an denied policy file")
 			}
 		} else {
 			param.PolicyConfigFilePaths = append(param.PolicyConfigFilePaths, policyFile)
